Refactored into multiple files, delete old version
