+++ enemy.py
@@ -1,0 +1,22 @@
+import pygame as pg
+import random
+from variables import enemy_images, BLACK, WIDTH, HEIGHT
+class Enemy(pg.sprite.Sprite):
+    def __init__(self):
+        pg.sprite.Sprite.__init__(self)
+        self.image_orig = random.choice(enemy_images)
+        self.image_orig.set_colorkey(BLACK)
+        self.image = self.image_orig.copy()
+        self.rect = self.image.get_rect()
+        self.rect.x = random.randrange(WIDTH - self.rect.width)
+        self.rect.bottom = random.randrange(-80, -20)
+        self.speedy = random.randrange(2, 10)
+        self.speedx = random.randrange(-2, 2)
+
+    def update(self):
+        self.rect.x += self.speedx
+        self.rect.y += self.speedy
+        if self.rect.top > HEIGHT + 10 or self.rect.left < -100 or self.rect.right > WIDTH + 100:
+            self.rect.x = random.randrange(WIDTH - self.rect.width)
+            self.rect.y = random.randrange(-100, -10)
+            self.speedy = random.randrange(2, 10)
